Adding deposit, withdraw, deleteAccount, doesAccountExists methods.

+++ src/main/java/banking/service/AccountServiceImpl.java
@@ -7,6 +7,8 @@ import java.util.Map;
 
 import javax.transaction.TransactionManager;
 
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,7 +29,6 @@ public class AccountServiceImpl implements AccountService{
 	
 	private Logger logger = LoggerFactory.getLogger(AccountServiceImpl.class);
 	
-	
 	@Autowired
 	private AccountDao accountDao;
 
@@ -45,6 +46,7 @@ public class AccountServiceImpl implements AccountService{
 	
 	
 	public Map<Integer, Account> showAllAccounts() {
+	
 		Map<Integer,Account> accountMap = accountDao.showAllAccounts();
 		Iterator it= accountMap.entrySet().iterator();
 		while(it.hasNext()){
@@ -56,8 +58,8 @@ public class AccountServiceImpl implements AccountService{
  		return null;
 	}
 	
-	
 	public String getSpecificAccount(Integer number){
+		if(doesAccountExists(number) != true){return null;}
 		Account account = accountDao.getSpeicificAccount(number);
 		showAccountDetails(account);
 		return null;
@@ -65,23 +67,28 @@ public class AccountServiceImpl implements AccountService{
 	
 
 	public double deposit(Integer number, Double deposit)  {
+		if(doesAccountExists(number) != true){return 0;}
 		accountDao.deposit(number, deposit);
 		return 0;
 	}
 
 	public double withdraw(Integer number, Double wd)  {
+		if(doesAccountExists(number) != true){return 0;}
 		accountDao.withdraw(number, wd);
 		return 0;
 	}
 
 	public boolean deleteAccount(Integer number) {
+		if(doesAccountExists(number) != true){return false;}
 		accountDao.deleteAccount(number);
 		return false;
 	}
 
 	public boolean doesAccountExists(Integer number)  {
-		accountDao.doesAccountExists(number);
-		return false;
+		boolean accountBool =accountDao.doesAccountExists(number);
+		if(accountBool==false){logger.info("doesAccountexists false, account number : " +number );
+		}else{logger.info("doesAccountexists ture, ccount number : " +number );}
+		return accountBool;
 	}
 	
 	public void showAccountDetails(Account account){
